requirements.txt
Django
djangorestframework
django-environ
django-model-utils
django-storages[azure]
pillow
whitenoise
django-cors-headers
mysqlclient
gunicorn
